Fin de passage des classes vers les entity
