+++ core/class/interactDef.class.php
@@ -22,577 +22,65 @@ use NextDom\Managers\InteractDefManager;
 
 require_once __DIR__ . '/../../core/php/core.inc.php';
 
-class interactDef {
-    /*     * *************************Attributs****************************** */
-
-    private $id;
-    private $name;
-    private $filtres;
-    private $query;
-    private $reply;
-    private $person;
-    private $options;
-    private $enable;
-    private $group;
-    private $actions;
-    private $_changed = false;
-
-    /*     * ***********************Méthodes statiques*************************** */
-
-    public static function byId($_id) {
+class interactDef extends \NextDom\Model\Entity\InteractDef
+{
+    public static function byId($_id)
+    {
         return InteractDefManager::byId($_id);
     }
 
-    public static function all($_group = '') {
+    public static function all($_group = '')
+    {
         return InteractDefManager::all($_group);
     }
 
-    public static function listGroup($_group = null) {
+    public static function listGroup($_group = null)
+    {
         return InteractDefManager::listGroup($_group);
     }
 
-    public static function generateTextVariant($_text) {
+    public static function generateTextVariant($_text)
+    {
         return InteractDefManager::generateTextVariant($_text);
     }
 
-    public static function searchByQuery($_query) {
+    public static function searchByQuery($_query)
+    {
         return InteractDefManager::searchByUse($_query);
     }
 
-    public static function regenerateInteract() {
+    public static function regenerateInteract()
+    {
         InteractDefManager::regenerateInteract();
     }
 
-    public static function getTagFromQuery($_def, $_query) {
+    public static function getTagFromQuery($_def, $_query)
+    {
         return InteractDefManager::getTagFromQuery($_def, $_query);
     }
 
-    public static function sanitizeQuery($_query) {
+    public static function sanitizeQuery($_query)
+    {
         return InteractDefManager::sanitizeQuery($_query);
     }
 
-    public static function deadCmd() {
+    public static function deadCmd()
+    {
         return InteractDefManager::deadCmd();
     }
 
-    public static function cleanInteract() {
+    public static function cleanInteract()
+    {
         return InteractDefManager::cleanInteract();
     }
 
-    public static function searchByUse($_search) {
+    public static function searchByUse($_search)
+    {
         return InteractDefManager::searchByUse($_search);
     }
 
-    /*     * *********************Méthodes d'instance************************* */
-
-    public function checkQuery($_query) {
-        if ($this->getOptions('allowSyntaxCheck', 1) == 1) {
-            $exclude_regexp = "/l'(z|r|t|p|q|s|d|f|g|j|k|l|m|w|x|c|v|b|n|y| )|( |^)la (a|e|u|i|o)|( |^)le (a|e|u|i|o)|( |^)du (a|e|u|i|o)/i";
-            if (preg_match($exclude_regexp, $_query)) {
-                return false;
-            }
-            $disallow = array(
-                'le salle',
-                'le chambre',
-                'la dressing',
-                'la salon',
-                'le cuisine',
-                'la jours',
-                'la total',
-                '(le|la) dehors',
-                'la balcon',
-                'du chambre',
-                'du salle',
-                'du cuisine',
-                'la homecinema',
-                'la led',
-                'le led',
-                'la pc',
-                'la sol',
-                'la conseil',
-                '(la|les) lave\-vaisselle',
-                '(la|les) lave\-linge',
-                'la sonos',
-                '(la|le) humidité',
-                'la genre',
-                'la résumé',
-                'le bouton',
-                'la status',
-                'la volume',
-                'le piste',
-                'le consommation',
-                'le position',
-                'le puissance',
-                'le luminosité',
-                'le température',
-                '(la|les) micro\-onde',
-                'la mirroir',
-                'la lapin',
-                'la greenmomit',
-                'le prise',
-                'le frigo',
-                'le (petite | )lumière',
-                'la boutton',
-                'la sommeil',
-                'la temps',
-                'la poids',
-                '(la|les) heartbeat',
-                '(la|le) heure',
-                'la nombre',
-                'la coût',
-                'la titre',
-                'la type',
-                'la demain',
-                'la pas',
-                'la démarré',
-                'la relai',
-                '(la|le) vacance',
-                'la coucher',
-                'la lever',
-                'la kodi',
-                'la frigo',
-                'la citronier',
-                'la basilique',
-                'la plante',
-                'la mouvement',
-                'la mode',
-                'la statut',
-                'la dns',
-                'la thym',
-                'lumière cuisine',
-                'lumière salon',
-                'lumière chambre',
-                'lumière salle de bain',
-                'la thumbnail',
-                'la bouton',
-                'la co',
-                'la co2',
-                'la répéter',
-                '(fait-il|combien) chambre',
-                '(fait-il|combien) salon',
-                '(fait-il|combien) cuisine',
-                '(fait-il|combien) salle',
-                '(fait-il|combien) entrée',
-                '(fait-il|combien) balcon',
-                '(fait-il|combien) appartement',
-                'dans le balcon',
-                'le calorie',
-                'le chansons',
-                'le charge',
-                'le demain',
-                'le démarré',
-                'le direction',
-                'le distance',
-                'le masse',
-                'le mémoire',
-                'le pr(é|e)sence',
-                'le répéter',
-                'le taille',
-                'le fumée',
-                'le pression',
-                'le vitesse',
-                'le condition',
-                'les pc',
-                'la tetris',
-                'le bougies',
-                'le myfox',
-                'les homecinema',
-                'les kodi',
-                'les appartement',
-                'le maison',
-                'du maison',
-                'le buanderie',
-                'du buanderie',
-                'la bureau',
-                'de salon',
-                'de maison',
-                'de chambre',
-                'de cuisine',
-                'de espace',
-                'de salle de bain',
-                '(dans|quelqu\'un) entr(é|e)e',
-            );
-            if (preg_match('/( |^)' . implode('( |$)|( |^)', $disallow) . '( |$)/i', $_query)) {
-                return false;
-            }
-        }
-        if ($this->getOptions('exclude_regexp') != '' && preg_match($this->getOptions('exclude_regexp'), $_query)) {
-            return false;
-        }
-        if (config::byKey('interact::regexpExcludGlobal') != '' && preg_match(config::byKey('interact::regexpExcludGlobal'), $_query)) {
-            return false;
-        }
-        return true;
-    }
-
-    public function selectReply() {
-        $replies = self::generateTextVariant($this->getReply());
-        $random = rand(0, count($replies) - 1);
-        return $replies[$random];
-    }
-
-    public function preInsert() {
-        if ($this->getReply() == '') {
-            $this->setReply('#valeur#');
-        }
-        $this->setEnable(1);
-    }
-
-    public function preSave() {
-        if ($this->getOptions('allowSyntaxCheck') === '') {
-            $this->setOptions('allowSyntaxCheck', 1);
-        }
-        if ($this->getFiltres('eqLogic_id') == '') {
-            $this->setFiltres('eqLogic_id', 'all');
-        }
-    }
-
-    public function save() {
-        if ($this->getQuery() == '') {
-            throw new Exception(__('La commande (demande) ne peut pas être vide', __FILE__));
-        }
-        DB::save($this);
-        return true;
-    }
-
-    public function postSave() {
-        $queries = $this->generateQueryVariant();
-        interactQuery::removeByInteractDefId($this->getId());
-        if ($this->getEnable()) {
-            DB::beginTransaction();
-            foreach ($queries as $query) {
-                $query['query'] = self::sanitizeQuery($query['query']);
-                if (trim($query['query']) == '') {
-                    continue;
-                }
-                if (!$this->checkQuery($query['query'])) {
-                    continue;
-                }
-                $interactQuery = new interactQuery();
-                $interactQuery->setInteractDef_id($this->getId());
-                $interactQuery->setQuery($query['query']);
-                $interactQuery->setActions('cmd', $query['cmd']);
-                $interactQuery->save();
-            }
-            DB::commit();
-        }
-        self::cleanInteract();
-    }
-
-    public function remove() {
-        DB::remove($this);
-    }
-
-    public function preRemove() {
-        interactQuery::removeByInteractDefId($this->getId());
-    }
-
-    public function postRemove() {
-        self::cleanInteract();
-    }
-
-    public function generateQueryVariant() {
-        $inputs = self::generateTextVariant($this->getQuery());
-        $return = array();
-        $object_filter = $this->getFiltres('object');
-        $type_filter = $this->getFiltres('type');
-        $subtype_filter = $this->getFiltres('subtype');
-        $unite_filter = $this->getFiltres('unite');
-        $plugin_filter = $this->getFiltres('plugin');
-        $visible_filter = $this->getFiltres('visible');
-        $category_filter = $this->getFiltres('category');
-        foreach ($inputs as $input) {
-            preg_match_all("/#(.*?)#/", $input, $matches);
-            $matches = $matches[1];
-            if (in_array('commande', $matches) || (in_array('objet', $matches) || in_array('equipement', $matches))) {
-                foreach (jeeObject::all() as $object) {
-                    if (isset($object_filter[$object->getId()]) && $object_filter[$object->getId()] == 0) {
-                        continue;
-                    }
-                    if (isset($visible_filter['object']) && $visible_filter['object'] == 1 && $object->getIsVisible() != 1) {
-                        continue;
-                    }
-                    foreach ($object->getEqLogic() as $eqLogic) {
-                        if ($this->getFiltres('eqLogic_id', 'all') != 'all' && $eqLogic->getId() != $this->getFiltres('eqLogic_id')) {
-                            continue;
-                        }
-                        if (isset($plugin_filter[$eqLogic->getEqType_name()]) && $plugin_filter[$eqLogic->getEqType_name()] == 0) {
-                            continue;
-                        }
-                        if (isset($visible_filter['eqLogic']) && $visible_filter['eqLogic'] == 1 && $eqLogic->getIsVisible() != 1) {
-                            continue;
-                        }
-
-                        $category_ok = true;
-                        if (is_array($category_filter)) {
-                            $category_ok = false;
-                            foreach ($category_filter as $category => $value) {
-                                if ($value == 1) {
-                                    if ($eqLogic->getCategory($category) == 1) {
-                                        $category_ok = true;
-                                        break;
-                                    }
-                                    if ($category == 'noCategory' && $eqLogic->getPrimaryCategory() == '') {
-                                        $category_ok = true;
-                                        break;
-                                    }
-                                }
-                            }
-                        }
-                        if (!$category_ok) {
-                            continue;
-                        }
-                        foreach ($eqLogic->getCmd() as $cmd) {
-                            if (isset($visible_filter['cmd']) && $visible_filter['cmd'] == 1 && $cmd->getIsVisible() != 1) {
-                                continue;
-                            }
-                            if (isset($subtype_filter[$cmd->getSubType()]) && $subtype_filter[$cmd->getSubType()] == 0) {
-                                continue;
-                            }
-                            if (isset($type_filter[$cmd->getType()]) && $type_filter[$cmd->getType()] == 0) {
-                                continue;
-                            }
-                            if ($cmd->getUnite() == '') {
-                                if (isset($unite_filter['none']) && $unite_filter['none'] == 0) {
-                                    continue;
-                                }
-                            } else {
-                                if (isset($unite_filter[$cmd->getUnite()]) && $unite_filter[$cmd->getUnite()] == 0) {
-                                    continue;
-                                }
-                            }
-
-                            $replace = array(
-                                '#objet#' => strtolower($object->getName()),
-                                '#commande#' => strtolower($cmd->getName()),
-                                '#equipement#' => strtolower($eqLogic->getName()),
-                            );
-                            $options = array();
-                            if ($cmd->getType() == 'action') {
-                                if ($cmd->getSubtype() == 'color') {
-                                    $options['color'] = '#color#';
-                                }
-                                if ($cmd->getSubtype() == 'slider') {
-                                    $options['slider'] = '#slider#';
-                                }
-                                if ($cmd->getSubtype() == 'message') {
-                                    $options['message'] = '#message#';
-                                    $options['title'] = '#title#';
-                                }
-                            }
-                            $query = str_replace(array_keys($replace), $replace, $input);
-                            $return[$query] = array(
-                                'query' => $query,
-                                'cmd' => array(array('cmd' => '#' . $cmd->getId() . '#', 'options' => $options)),
-
-                            );
-                        }
-                    }
-                }
-            }
-        }
-
-        if (count($return) == 0) {
-            foreach ($inputs as $input) {
-                $return[] = array(
-                    'query' => $input,
-                    'cmd' => $this->getActions('cmd'),
-                );
-            }
-        }
-        if ($this->getOptions('synonymes') != '') {
-            $synonymes = array();
-            foreach (explode('|', $this->getOptions('synonymes')) as $value) {
-                $values = explode('=', $value);
-                if (count($values) != 2) {
-                    continue;
-                }
-                $synonymes[self::sanitizeQuery($values[0])] = explode(',', self::sanitizeQuery($values[1]));
-            }
-            $result = array();
-            foreach ($return as $query) {
-                $results = self::generateSynonymeVariante(self::sanitizeQuery($query['query']), $synonymes);
-                if (count($results) == 0) {
-                    continue;
-                }
-                foreach ($results as $result) {
-                    $query_info = $query;
-                    $query_info['query'] = $result;
-                    $return[$result] = $query_info;
-                }
-            }
-        }
-        return $return;
-    }
-
-    public static function generateSynonymeVariante($_text, $_synonymes, $_deep = 0) {
-        $return = array();
-        if (count($_synonymes) == 0) {
-            return $return;
-        }
-        if ($_deep > 10) {
-            return $return;
-        }
-        $_deep++;
-        foreach ($_synonymes as $replace => $values) {
-            foreach ($values as $value) {
-                $result = @preg_replace('/\b' . $replace . '\b/iu', $value, $_text);
-                if ($result != $_text) {
-                    $synonymes = $_synonymes;
-                    unset($synonymes[$replace]);
-                    $return = array_merge($return, self::generateSynonymeVariante($result, $synonymes, $_deep));
-                    $return[] = $result;
-                }
-            }
-        }
-        return $return;
-    }
-
-    public function getLinkToConfiguration() {
-        return 'index.php?v=d&p=interact&id=' . $this->getId();
-    }
-
-    public function getHumanName() {
-        if ($this->getName() != '') {
-            return $this->getName();
-        }
-        return $this->getQuery();
-    }
-
-    public function getLinkData(&$_data = array('node' => array(), 'link' => array()), $_level = 0, $_drill = 3) {
-        if (isset($_data['node']['interactDef' . $this->getId()])) {
-            return;
-        }
-        $_level++;
-        if ($_level > $_drill) {
-            return $_data;
-        }
-        $icon = findCodeIcon('fa-comments-o');
-        $_data['node']['interactDef' . $this->getId()] = array(
-            'id' => 'interactDef' . $this->getId(),
-            'name' => substr($this->getHumanName(), 0, 20),
-            'icon' => $icon['icon'],
-            'fontfamily' => $icon['fontfamily'],
-            'fontsize' => '1.5em',
-            'fontweight' => ($_level == 1) ? 'bold' : 'normal',
-            'texty' => -14,
-            'textx' => 0,
-            'title' => $this->getHumanName(),
-            'url' => 'index.php?v=d&p=interact&id=' . $this->getId(),
-        );
-    }
-
-/*     * **********************Getteur Setteur*************************** */
-
-    public function getId() {
-        return $this->id;
-    }
-
-    public function setId($_id) {
-        $this->_changed = utils::attrChanged($this->_changed,$this->id,$_id);
-        $this->id = $_id;
-        return $this;
-    }
-
-    public function getQuery() {
-        return $this->query;
-    }
-
-    public function setQuery($_query) {
-        $this->_changed = utils::attrChanged($this->_changed,$this->query,$_query);
-        $this->query = $_query;
-        return $this;
-    }
-
-    public function getReply() {
-        return $this->reply;
-    }
-
-    public function setReply($_reply) {
-        $this->_changed = utils::attrChanged($this->_changed,$this->reply,$_reply);
-        $this->reply = $_reply;
-        return $this;
-    }
-
-    public function getPerson() {
-        return $this->person;
-    }
-
-    public function setPerson($_person) {
-        $this->_changed = utils::attrChanged($this->_changed,$this->person,$_person);
-        $this->person = $_person;
-        return $this;
-    }
-
-    public function getOptions($_key = '', $_default = '') {
-        return utils::getJsonAttr($this->options, $_key, $_default);
-    }
-
-    public function setOptions($_key, $_value) {
-        $options = utils::setJsonAttr($this->options, $_key, $_value);
-        $this->_changed = utils::attrChanged($this->_changed,$this->options,$options);
-        $this->options = $options;
-        return $this;
-    }
-
-    public function getFiltres($_key = '', $_default = '') {
-        return utils::getJsonAttr($this->filtres, $_key, $_default);
-    }
-
-    public function setFiltres($_key, $_value) {
-        $filtres = utils::setJsonAttr($this->filtres, $_key, $_value);
-        $this->_changed = utils::attrChanged($this->_changed,$this->filtres,$filtres);
-        $this->filtres = $filtres;
-        return $this;
-    }
-
-    public function getEnable() {
-        return $this->enable;
-    }
-
-    public function setEnable($_enable) {
-        $this->_changed = utils::attrChanged($this->_changed,$this->enable,$_enable);
-        $this->enable = $_enable;
-        return $this;
-    }
-
-    public function getName() {
-        return $this->name;
-    }
-
-    public function setName($_name) {
-        $this->_changed = utils::attrChanged($this->_changed,$this->name,$_name);
-        $this->name = $_name;
-        return $this;
-    }
-
-    public function getGroup() {
-        return $this->group;
-    }
-
-    public function setGroup($_group) {
-        $this->_changed = utils::attrChanged($this->_changed,$this->group,$_group);
-        $this->group = $_group;
-        return $this;
-    }
-
-    public function getActions($_key = '', $_default = '') {
-        return utils::getJsonAttr($this->actions, $_key, $_default);
-    }
-
-    public function setActions($_key, $_value) {
-        $actions = utils::setJsonAttr($this->actions, $_key, $_value);
-        $this->_changed = utils::attrChanged($this->_changed,$this->actions,$actions);
-        $this->actions = $actions;
-        return $this;
-    }
-
-    public function getChanged() {
-        return $this->_changed;
-    }
-
-    public function setChanged($_changed) {
-        $this->_changed = $_changed;
-        return $this;
+    public static function generateSynonymeVariante($_text, $_synonymes, $_deep = 0)
+    {
+        return InteractDefManager::generateSynonymeVariante($_text, $_synonymes, $_deep);
     }
 }
