+++ core/class/note.class.php
@@ -17,90 +17,18 @@
 */
 
 /* * ***************************Includes********************************* */
+
+use NextDom\Managers\NoteManager;
+
 require_once __DIR__ . '/../../core/php/core.inc.php';
 
-class note {
-    /*     * *************************Attributs****************************** */
-    
-    private $id;
-    private $name;
-    private $text;
-    private $_changed = false;
-    
-    /*     * ***********************Methode static*************************** */
-    
+class note extends \NextDom\Model\Entity\Note {
+
     public static function byId($_id) {
-        $values = array(
-            'id' => $_id,
-        );
-        $sql = 'SELECT ' . DB::buildField(__CLASS__) . '
-        FROM note
-        WHERE id=:id';
-        return DB::Prepare($sql, $values, DB::FETCH_TYPE_ROW, PDO::FETCH_CLASS, __CLASS__);
+        return NoteManager::byId($_id);
     }
     
     public static function all() {
-        $sql = 'SELECT ' . DB::buildField(__CLASS__) . '
-        FROM note
-        ORDER BY name';
-        return DB::Prepare($sql, array(), DB::FETCH_TYPE_ALL, PDO::FETCH_CLASS, __CLASS__);
+        return NoteManager::all();
     }
-    /*     * *********************Methode d'instance************************* */
-    
-    public function preSave() {
-        if (trim($this->getName()) == '') {
-            throw new Exception(__('Le nom de la note ne peut être vide', __FILE__));
-        }
-    }
-    
-    public function save() {
-        DB::save($this);
-        return true;
-    }
-    
-    public function remove() {
-        DB::remove($this);
-    }
-    
-    /*     * **********************Getteur Setteur*************************** */
-    
-    public function getId() {
-        return $this->id;
-    }
-    
-    public function getName() {
-        return $this->name;
-    }
-    
-    public function getText() {
-        return $this->text;
-    }
-    
-    public function setId($_id) {
-        $this->_changed = utils::attrChanged($this->_changed,$this->id,$_id);
-        $this->id = $_id;
-        return $this;
-    }
-    
-    public function setName($_name) {
-        $this->_changed = utils::attrChanged($this->_changed,$this->name,$_name);
-        $this->name = $_name;
-        return $this;
-    }
-    
-    public function setText($_text) {
-        $this->_changed = utils::attrChanged($this->_changed,$this->text,$_text);
-        $this->text = $_text;
-        return $this;
-    }
-    
-    public function getChanged() {
-        return $this->_changed;
-    }
-    
-    public function setChanged($_changed) {
-        $this->_changed = $_changed;
-        return $this;
-    }
-    
 }
