+++ src/Managers/NoteManager.php
@@ -1,0 +1,64 @@
+<?php
+/*
+* This file is part of the NextDom software (https://github.com/NextDom or http://nextdom.github.io).
+* Copyright (c) 2018 NextDom.
+*
+* This program is free software: you can redistribute it and/or modify
+* it under the terms of the GNU General Public License as published by
+* the Free Software Foundation, version 2.
+*
+* This program is distributed in the hope that it will be useful, but
+* WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+* General Public License for more details.
+*
+* You should have received a copy of the GNU General Public License
+* along with this program. If not, see <http://www.gnu.org/licenses/>.
+*/
+
+/* This file is part of Jeedom.
+ *
+ * Jeedom is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Jeedom is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Jeedom. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+namespace NextDom\Managers;
+
+use NextDom\Helpers\SystemHelper;
+use NextDom\Model\Entity\Listener;
+use NextDom\Model\Entity\Note;
+
+require_once NEXTDOM_ROOT . '/core/class/cache.class.php';
+
+class NoteManager
+{
+    const CLASS_NAME = Note::class;
+    const DB_CLASS_NAME = '`note`';
+
+    public static function byId($_id) {
+        $values = array(
+            'id' => $_id,
+        );
+        $sql = 'SELECT ' . \DB::buildField(self::CLASS_NAME) . '
+                FROM ' . self::DB_CLASS_NAME . '
+                WHERE id = :id';
+        return \DB::Prepare($sql, $values, \DB::FETCH_TYPE_ROW, \PDO::FETCH_CLASS, self::CLASS_NAME);
+    }
+
+    public static function all() {
+        $sql = 'SELECT ' . \DB::buildField(self::CLASS_NAME) . '
+                FROM ' . self::DB_CLASS_NAME . '
+                ORDER BY name';
+        return \DB::Prepare($sql, array(), \DB::FETCH_TYPE_ALL, \PDO::FETCH_CLASS, self::CLASS_NAME);
+    }
+}
