+++ src/Model/Entity/Cron.php
@@ -449,7 +449,7 @@ class Cron
             $this->setState('stop');
             $this->setPID();
         } else {
-            LogHelper::add('cron', 'info', __('Arrêt de ', __FILE__) . $this->getClass() . '::' . $this->getFunction() . '(), PID : ' . $this->getPID());
+            LogHelper::add('cron', 'info', __('Arrêt de ') . $this->getClass() . '::' . $this->getFunction() . '(), PID : ' . $this->getPID());
             if ($this->getPID() > 0) {
                 SystemHelper::kill($this->getPID());
                 $retry = 0;
@@ -475,7 +475,7 @@ class Cron
                 if ($this->running()) {
                     $this->setState('error');
                     $this->setPID();
-                    throw new CoreException($this->getClass() . '::' . $this->getFunction() . __('() : Impossible d\'arrêter la tâche', __FILE__));
+                    throw new CoreException($this->getClass() . '::' . $this->getFunction() . __('() : Impossible d\'arrêter la tâche'));
                 }
             } else {
                 $this->setState('stop');
