+++ src/Model/Entity/History.php
@@ -17,6 +17,9 @@
 
 namespace NextDom\Model\Entity;
 
+use NextDom\Managers\CmdManager;
+use NextDom\Managers\HistoryManager;
+
 /**
  * History
  *
@@ -30,14 +33,14 @@ class History
      *
      * @ORM\Column(name="datetime", type="datetime", nullable=false)
      */
-    private $datetime;
+    protected $datetime;
 
     /**
      * @var string
      *
      * @ORM\Column(name="value", type="string", length=127, nullable=true)
      */
-    private $value;
+    protected $value;
 
     /**
      * @var integer
@@ -46,7 +49,7 @@ class History
      * @ORM\Id
      * @ORM\GeneratedValue(strategy="IDENTITY")
      */
-    private $id;
+    protected $id;
 
     /**
      * @var \NextDom\Model\Entity\Cmd
@@ -56,11 +59,102 @@ class History
      *   @ORM\JoinColumn(name="cmd_id", referencedColumnName="id")
      * })
      */
-    private $cmd;
+    protected $cmd_id;
+    protected $_tableName = 'history';
 
-    public function getDatetime(): \DateTime
+    public function save($_cmd = null, $_direct = false)
     {
-        return $this->datetime;
+        global $NEXTDOM_INTERNAL_CONFIG;
+        if ($_cmd === null) {
+            $cmd = $this->getCmd();
+            if (!is_object($cmd)) {
+                HistoryManager::emptyHistory($this->getCmd_id());
+                return;
+            }
+        } else {
+            $cmd = $_cmd;
+        }
+        if ($this->getDatetime() == '') {
+            $this->setDatetime(date('Y-m-d H:i:s'));
+        }
+        if ($cmd->getConfiguration('historizeRound') !== '' && is_numeric($cmd->getConfiguration('historizeRound')) && $cmd->getConfiguration('historizeRound') >= 0 && $this->getValue() !== null) {
+            $this->setValue(round($this->getValue(), $cmd->getConfiguration('historizeRound')));
+        }
+        if ($NEXTDOM_INTERNAL_CONFIG['cmd']['type']['info']['subtype'][$cmd->getSubType()]['isHistorized']['canBeSmooth'] && $cmd->getConfiguration('historizeMode', 'avg') != 'none' && $this->getValue() !== null && $_direct === false) {
+            if ($this->getTableName() == 'history') {
+                $time = strtotime($this->getDatetime());
+                $time -= $time % 300;
+                $this->setDatetime(date('Y-m-d H:i:s', $time));
+                if ($this->getValue() === 0) {
+                    $values = array(
+                        'cmd_id' => $this->getCmd_id(),
+                        'datetime' => date('Y-m-d H:i:00', strtotime($this->getDatetime()) + 300),
+                        'value' => $this->getValue(),
+                    );
+                    $sql = 'REPLACE INTO history
+                    SET cmd_id=:cmd_id,
+                    `datetime`=:datetime,
+                    value=:value';
+                    \DB::Prepare($sql, $values, \DB::FETCH_TYPE_ROW);
+                    return;
+                }
+                $values = array(
+                    'cmd_id' => $this->getCmd_id(),
+                    'datetime' => $this->getDatetime(),
+                );
+                $sql = 'SELECT `value`
+                FROM history
+                WHERE cmd_id=:cmd_id
+                AND `datetime`=:datetime';
+                $result = \DB::Prepare($sql, $values, \DB::FETCH_TYPE_ROW);
+                if ($result !== false) {
+                    switch ($cmd->getConfiguration('historizeMode', 'avg')) {
+                        case 'avg':
+                            $this->setValue(($result['value'] + $this->getValue()) / 2);
+                            break;
+                        case 'min':
+                            $this->setValue(min($result['value'], $this->getValue()));
+                            break;
+                        case 'max':
+                            $this->setValue(max($result['value'], $this->getValue()));
+                            break;
+                    }
+                    if ($result['value'] === $this->getValue()) {
+                        return;
+                    }
+                }
+            } else {
+                $this->setDatetime(date('Y-m-d H:00:00', strtotime($this->getDatetime())));
+            }
+        }
+        $values = array(
+            'cmd_id' => $this->getCmd_id(),
+            'datetime' => $this->getDatetime(),
+            'value' => $this->getValue(),
+        );
+        if ($values['value'] === '') {
+            $values['value'] = null;
+        }
+        $sql = 'REPLACE INTO ' . $this->getTableName() . '
+        SET cmd_id=:cmd_id,
+        `datetime`=:datetime,
+        value=:value';
+        \DB::Prepare($sql, $values, \DB::FETCH_TYPE_ROW);
+    }
+
+    public function remove()
+    {
+        \DB::remove($this);
+    }
+
+    public function getCmd_id()
+    {
+        return $this->cmd_id;
+    }
+
+    public function getCmd()
+    {
+        return CmdManager::byId($this->cmd_id);
     }
 
     public function getValue()
@@ -68,19 +162,25 @@ class History
         return $this->value;
     }
 
-    public function getId()
+    public function getDatetime()
     {
-        return $this->id;
+        return $this->datetime;
     }
 
-    public function getCmd(): Cmd
+    public function getTableName()
     {
-        return $this->cmd;
+        return $this->_tableName;
     }
 
-    public function setDatetime(\DateTime $datetime)
+    public function setTableName($_tableName)
     {
-        $this->datetime = $datetime;
+        $this->_tableName = $_tableName;
+        return $this;
+    }
+
+    public function setCmd_id($cmd_id)
+    {
+        $this->cmd_id = $cmd_id;
         return $this;
     }
 
@@ -90,18 +190,10 @@ class History
         return $this;
     }
 
-    public function setId($id)
+    public function setDatetime($datetime)
     {
-        $this->id = $id;
+        $this->datetime = $datetime;
         return $this;
     }
-
-    public function setCmd(Cmd $cmd)
-    {
-        $this->cmd = $cmd;
-        return $this;
-    }
-
-
 }
