+++ src/Model/Entity/Listener.php
@@ -17,6 +17,11 @@
 
 namespace NextDom\Model\Entity;
 
+use NextDom\Helpers\LogHelper;
+use NextDom\Helpers\SystemHelper;
+use NextDom\Helpers\Utils;
+use NextDom\Managers\ListenerManager;
+
 /**
  * Listener
  *
@@ -31,28 +36,28 @@ class Listener
      *
      * @ORM\Column(name="class", type="string", length=127, nullable=true)
      */
-    private $class;
+    protected $class;
 
     /**
      * @var string
      *
      * @ORM\Column(name="function", type="string", length=127, nullable=true)
      */
-    private $function;
+    protected $function;
 
     /**
      * @var string
      *
      * @ORM\Column(name="event", type="string", length=255, nullable=true)
      */
-    private $event;
+    protected $event;
 
     /**
      * @var string
      *
      * @ORM\Column(name="option", type="text", length=65535, nullable=true)
      */
-    private $option;
+    protected $option;
 
     /**
      * @var integer
@@ -61,7 +66,114 @@ class Listener
      * @ORM\Id
      * @ORM\GeneratedValue(strategy="IDENTITY")
      */
-    private $id;
+    protected $id;
+
+    protected $_changed = false;
+
+    public function run($_event, $_value, $_datetime = null)
+    {
+        $option = array();
+        if (count($this->getOption()) > 0) {
+            $option = $this->getOption();
+        }
+        if (isset($option['background']) && $option['background'] == false) {
+            $this->execute($_event, $_value);
+        } else {
+            $cmd = NEXTDOM_ROOT . '/core/php/jeeListener.php';
+            $cmd .= ' listener_id=' . $this->getId() . ' event_id=' . $_event . ' "value=' . escapeshellarg($_value) . '"';
+            if ($_datetime !== null) {
+                $cmd .= ' "datetime=' . escapeshellarg($_datetime) . '"';
+            }
+            SystemHelper::php($cmd . ' >> ' . LogHelper::getPathToLog('listener_execution') . ' 2>&1 &');
+        }
+    }
+
+    public function execute($_event, $_value, $_datetime = '')
+    {
+        try {
+            $option = array();
+            if (count($this->getOption()) > 0) {
+                $option = $this->getOption();
+            }
+            $option['event_id'] = $_event;
+            $option['value'] = $_value;
+            $option['datetime'] = $_datetime;
+            $option['listener_id'] = $this->getId();
+            if ($this->getClass() != '') {
+                $class = $this->getClass();
+                $function = $this->getFunction();
+                if (class_exists($class) && method_exists($class, $function)) {
+                    $class::$function($option);
+                } else {
+                    LogHelper::add('listener', 'debug', __('[Erreur] Classe ou fonction non trouvée ') . $this->getName());
+                    $this->remove();
+                    return;
+                }
+            } else {
+                $function = $this->getFunction();
+                if (function_exists($function)) {
+                    $function($option);
+                } else {
+                    LogHelper::add('listener', 'error', __('[Erreur] Non trouvée ') . $this->getName());
+                    return;
+                }
+            }
+        } catch (\Exception $e) {
+            LogHelper::add(init('plugin_id', 'plugin'), 'error', $e->getMessage());
+        }
+    }
+
+    public function preSave()
+    {
+        if ($this->getFunction() == '') {
+            throw new \Exception(__('La fonction ne peut pas être vide'));
+        }
+    }
+
+    public function save($_once = false)
+    {
+        if ($_once) {
+            ListenerManager::removeByClassFunctionAndEvent($this->getClass(), $this->getFunction(), $this->event, $this->getOption());
+        }
+        \DB::save($this);
+        return true;
+    }
+
+    public function remove()
+    {
+        return \DB::remove($this);
+    }
+
+    public function emptyEvent()
+    {
+        $this->event = array();
+    }
+
+    public function addEvent($_id, $_type = 'cmd')
+    {
+        $event = $this->getEvent();
+        if (!is_array($event)) {
+            $event = array();
+        }
+        $id = '';
+        if ($_type == 'cmd') {
+            $id = str_replace('#', '', $_id);
+        }
+        if (!in_array('#' . $id . '#', $event)) {
+            $event[] = '#' . $id . '#';
+        }
+        $this->setEvent($event);
+    }
+
+    public function getId()
+    {
+        return $this->id;
+    }
+
+    public function getEvent()
+    {
+        return Utils::isJson($this->event, array());
+    }
 
     public function getClass()
     {
@@ -73,49 +185,70 @@ class Listener
         return $this->function;
     }
 
-    public function getEvent()
+    public function getOption($_key = '', $_default = '')
     {
-        return $this->event;
+        return Utils::getJsonAttr($this->option, $_key, $_default);
     }
 
-    public function getOption()
+    public function setId($_id)
     {
-        return $this->option;
-    }
-
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    public function setClass($class)
-    {
-        $this->class = $class;
+        $this->_changed = Utils::attrChanged($this->_changed, $this->id, $_id);
+        $this->id = $_id;
         return $this;
     }
 
-    public function setFunction($function)
+    public function setEvent($_event)
     {
-        $this->function = $function;
-        return $this;
-    }
-
-    public function setEvent($event)
-    {
+        $event = json_encode($_event, JSON_UNESCAPED_UNICODE);
+        $this->_changed = Utils::attrChanged($this->_changed, $this->event, $_event);
         $this->event = $event;
         return $this;
     }
 
-    public function setOption($option)
+    public function setClass($_class)
     {
+        $this->_changed = Utils::attrChanged($this->_changed, $this->class, $_class);
+        $this->class = $_class;
+        return $this;
+    }
+
+    public function setFunction($_function)
+    {
+        $this->_changed = Utils::attrChanged($this->_changed, $this->function, $_function);
+        $this->function = $_function;
+        return $this;
+    }
+
+    public function setOption($_key, $_value = '')
+    {
+        $option = Utils::setJsonAttr($this->option, $_key, $_value);
+        $this->_changed = Utils::attrChanged($this->_changed, $this->option, $option);
         $this->option = $option;
         return $this;
     }
 
-    public function setId($id)
+    public function getChanged()
     {
-        $this->id = $id;
+        return $this->_changed;
+    }
+
+    public function setChanged($_changed)
+    {
+        $this->_changed = $_changed;
         return $this;
+    }
+
+    public function getName()
+    {
+        if ($this->getClass() != '') {
+            return $this->getClass() . '::' . $this->getFunction() . '()';
+        }
+        return $this->getFunction() . '()';
+    }
+
+    public function getTableName()
+    {
+        return 'listener';
     }
 
 }
