+++ src/Model/Entity/ScenarioExpression.php
@@ -22,6 +22,7 @@ use NextDom\Enums\ScenarioExpressionTypeEnum;
 use NextDom\Exceptions\CoreException;
 use NextDom\Helpers\NetworkHelper;
 use NextDom\Helpers\NextDomHelper;
+use NextDom\Helpers\ReportHelper;
 use NextDom\Helpers\SystemHelper;
 use NextDom\Helpers\Utils;
 use NextDom\Managers\CacheManager;
@@ -193,7 +194,7 @@ class ScenarioExpression
         return null;
     }
 
-    private function executeAction(&$scenario, $options)
+    protected function executeAction(&$scenario, $options)
     {
         switch ($this->getExpression()) {
             case ScenarioExpressionEnum::ICON:
@@ -266,7 +267,7 @@ class ScenarioExpression
     /**
      * @param Scenario $scenario
      */
-    private function executeActionIcon(&$scenario)
+    protected function executeActionIcon(&$scenario)
     {
         if ($scenario !== null) {
             $options = $this->getOptions();
@@ -276,7 +277,7 @@ class ScenarioExpression
         }
     }
 
-    private function executeActionWait(&$scenario, $options)
+    protected function executeActionWait(&$scenario, $options)
     {
         if (!isset($options['condition'])) {
             return null;
@@ -302,7 +303,7 @@ class ScenarioExpression
         $this->setLog($scenario, __('[Wait] Condition valide : ') . $expression . ' => ' . $result);
     }
 
-    private function executeActionSleep(&$scenario, $options)
+    protected function executeActionSleep(&$scenario, $options)
     {
         if (isset($options['duration'])) {
             try {
@@ -327,7 +328,7 @@ class ScenarioExpression
      * @param Scenario $scenario
      * @return null
      */
-    private function executeActionStop(&$scenario)
+    protected function executeActionStop(&$scenario)
     {
         if ($scenario !== null) {
             $this->setLog($scenario, __('Action stop'));
@@ -341,7 +342,7 @@ class ScenarioExpression
      * @param Scenario $scenario
      * @param $options
      */
-    private function executeActionLog(&$scenario, $options)
+    protected function executeActionLog(&$scenario, $options)
     {
         if ($scenario !== null) {
             $scenario->setLog('Log : ' . $options['message']);
@@ -353,7 +354,7 @@ class ScenarioExpression
      * @param $options
      * @throws CoreException
      */
-    private function executeActionEvent(&$scenario, $options)
+    protected function executeActionEvent(&$scenario, $options)
     {
         $cmd = CmdManager::byId(trim(str_replace('#', '', $options['cmd'])));
         if (!is_object($cmd)) {
@@ -363,25 +364,25 @@ class ScenarioExpression
         $cmd->event(NextDomHelper::evaluateExpression($options['value']));
     }
 
-    private function executeActionMessage(&$scenario, $options)
+    protected function executeActionMessage(&$scenario, $options)
     {
         \message::add('scenario', $options['message']);
         $this->setLog($scenario, __('Ajout du message suivant dans le centre de message : ') . $options['message']);
     }
 
-    private function executeActionAlert(&$scenario, $options)
+    protected function executeActionAlert(&$scenario, $options)
     {
         EventManager::add('nextdom::alert', $options);
         $this->setLog($scenario, __('Ajout de l\'alerte : ') . $options['message']);
     }
 
-    private function executeActionPopup(&$scenario, $options)
+    protected function executeActionPopup(&$scenario, $options)
     {
         EventManager::add('nextdom::alertPopup', $options['message']);
         $this->setLog($scenario, __('Affichage du popup : ') . $options['message']);
     }
 
-    private function executeActionEquipment(&$scenario)
+    protected function executeActionEquipment(&$scenario)
     {
         $eqLogic = EqLogicManager::byId(str_replace(array('#eqLogic', '#'), '', $this->getOptions('eqLogic')));
         if (!is_object($eqLogic)) {
@@ -411,7 +412,7 @@ class ScenarioExpression
         }
     }
 
-    private function executeActionGotoDesign(&$scenario, $options)
+    protected function executeActionGotoDesign(&$scenario, $options)
     {
         $this->setLog($scenario, __('Changement design : ') . $options['plan_id']);
         EventManager::add('nextdom::gotoplan', $options['plan_id']);
@@ -423,7 +424,7 @@ class ScenarioExpression
      * @throws CoreException
      * @throws \ReflectionException
      */
-    private function executeActionScenario(&$scenario)
+    protected function executeActionScenario(&$scenario)
     {
         if ($scenario !== null && $this->getOptions('scenario_id') == $scenario->getId()) {
             $actionScenario = &$scenario;
@@ -494,7 +495,7 @@ class ScenarioExpression
         return null;
     }
 
-    private function executeActionVariable(&$scenario, $options)
+    protected function executeActionVariable(&$scenario, $options)
     {
         $options['value'] = ScenarioExpressionManager::setTags($options['value'], $scenario);
         try {
@@ -521,14 +522,19 @@ class ScenarioExpression
      * @return null
      * @throws \Exception
      */
-    private function executeActionDeleteVariable(&$scenario, $options)
+    protected function executeActionDeleteVariable(&$scenario, $options)
     {
         $scenario->removeData($options['name']);
         $this->setLog($scenario, __('Suppression de la variable ') . $this->getOptions('name'));
         return null;
     }
 
-    private function executeActionAsk(&$scenario, $options)
+    /**
+     * @param Scenario $scenario
+     * @param $options
+     * @throws CoreException
+     */
+    protected function executeActionAsk(&$scenario, $options)
     {
         $dataStore = new \dataStore();
         $dataStore->setType('scenario');
@@ -584,7 +590,7 @@ class ScenarioExpression
     /**
      * @param Scenario $scenario
      */
-    private function executeActionNextDomPowerOff(&$scenario)
+    protected function executeActionNextDomPowerOff(&$scenario)
     {
         $this->setLog($scenario, __('Lancement de l\'arret de nextdom'));
         $scenario->persistLog();
@@ -595,7 +601,7 @@ class ScenarioExpression
      * @param Scenario $scenario
      * @param $options
      */
-    private function executeActionScenarioReturn(&$scenario, $options)
+    protected function executeActionScenarioReturn(&$scenario, $options)
     {
         $this->setLog($scenario, __('Demande de retour d\'information : ') . $options['message']);
         if ($scenario->getReturn() === true) {
@@ -609,7 +615,7 @@ class ScenarioExpression
      * @param Scenario $scenario
      * @throws \Exception
      */
-    private function executeActionRemoveInat(&$scenario)
+    protected function executeActionRemoveInat(&$scenario)
     {
         if ($scenario !== null) {
             $this->setLog($scenario, __('Suppression des blocs DANS et A programmés du scénario '));
@@ -624,7 +630,7 @@ class ScenarioExpression
         }
     }
 
-    private function executeActionReport(&$scenario, $options)
+    protected function executeActionReport(&$scenario, $options)
     {
         $cmd_parameters = array('files' => null);
         $this->setLog($scenario, __('Génération d\'un rapport de type ') . $options['type']);
@@ -662,7 +668,7 @@ class ScenarioExpression
             case 'eqAnalyse':
                 $url = NetworkHelper::getNetworkAccess('internal') . '/index.php?v=d&p=eqAnalyse&report=1';
                 $this->setLog($scenario, __('Génération du rapport ') . $url);
-                $cmd_parameters['files'] = array(\report::generate($url,'other',$options['export_type'], $options));
+                $cmd_parameters['files'] = array(ReportHelper::generate($url, 'other', $options['export_type'], $options));
                 $cmd_parameters['title'] = __('[' . ConfigManager::byKey('name') . '] Rapport équipement du ') . date('Y-m-d H:i:s');
                 $cmd_parameters['message'] = __('Veuillez trouver ci-joint le rapport équipement généré le ') . date('Y-m-d H:i:s');
                 break;
@@ -684,7 +690,7 @@ class ScenarioExpression
      * @param Scenario $scenario
      * @param $options
      */
-    private function executeActionTag(&$scenario, $options)
+    protected function executeActionTag(&$scenario, $options)
     {
         $tags = $scenario->getTags();
         $tags['#' . $options['name'] . '#'] = $options['value'];
@@ -698,7 +704,7 @@ class ScenarioExpression
      * @return mixed
      * @throws CoreException
      */
-    private function executeActionOthers(&$scenario, $options)
+    protected function executeActionOthers(&$scenario, $options)
     {
         $cmd = CmdManager::byId(str_replace('#', '', $this->getExpression()));
         if (is_object($cmd)) {
