+++ src/Model/Entity/User.php
@@ -91,13 +91,15 @@ class User
      */
     protected $id;
 
-    public function preInsert() {
+    public function preInsert()
+    {
         if (is_object(UserManager::byLogin($this->getLogin()))) {
             throw new \Exception(__('Ce nom d\'utilisateur est déja pris'));
         }
     }
 
-    public function preSave() {
+    public function preSave()
+    {
         if ($this->getLogin() == '') {
             throw new \Exception(__('Le nom d\'utilisateur ne peut pas être vide'));
         }
@@ -110,22 +112,26 @@ class User
         }
     }
 
-    public function save() {
+    public function save()
+    {
         return \DB::save($this);
     }
 
-    public function preRemove() {
+    public function preRemove()
+    {
         if (count(UserManager::byProfils('admin', true)) == 1 && $this->getProfils() == 'admin') {
             throw new \Exception(__('Vous ne pouvez supprimer le dernier administrateur'));
         }
     }
 
-    public function remove() {
+    public function remove()
+    {
         NextDomHelper::addRemoveHistory(array('id' => $this->getId(), 'name' => $this->getLogin(), 'date' => date('Y-m-d H:i:s'), 'type' => 'user'));
         return \DB::remove($this);
     }
 
-    public function refresh() {
+    public function refresh()
+    {
         \DB::refresh($this);
     }
 
@@ -133,11 +139,13 @@ class User
      *
      * @return boolean vrai si l'utilisateur est valide
      */
-    public function is_Connected() {
+    public function is_Connected()
+    {
         return (is_numeric($this->id) && $this->login != '');
     }
 
-    public function validateTwoFactorCode($_code) {
+    public function validateTwoFactorCode($_code)
+    {
         $google2fa = new Google2FA();
         return $google2fa->verifyKey($this->getOptions('twoFactorAuthentificationSecret'), $_code);
     }
@@ -145,70 +153,83 @@ class User
     /*     * **********************Getteur Setteur*************************** */
 
 
-    public function getId() {
+    public function getId()
+    {
         return $this->id;
     }
 
-    public function getLogin() {
+    public function getLogin()
+    {
         return $this->login;
     }
 
-    public function getPassword() {
+    public function getPassword()
+    {
         return $this->password;
     }
 
-    public function setId($_id) {
-        $this->_changed = Utils::attrChanged($this->_changed,$this->id,$_id);
+    public function setId($_id)
+    {
+        $this->_changed = Utils::attrChanged($this->_changed, $this->id, $_id);
         $this->id = $_id;
         return $this;
     }
 
-    public function setLogin($_login) {
-        $this->_changed = Utils::attrChanged($this->_changed,$this->login,$_login);
+    public function setLogin($_login)
+    {
+        $this->_changed = Utils::attrChanged($this->_changed, $this->login, $_login);
         $this->login = $_login;
         return $this;
     }
 
-    public function setPassword($_password) {
+    public function setPassword($_password)
+    {
         $_password = (!Utils::isSha512($_password)) ? Utils::sha512($_password) : $_password;
-        $this->_changed = Utils::attrChanged($this->_changed,$this->password,$_password);
+        $this->_changed = Utils::attrChanged($this->_changed, $this->password, $_password);
         $this->password = $_password;
         return $this;
     }
 
-    public function getOptions($_key = '', $_default = '') {
+    public function getOptions($_key = '', $_default = '')
+    {
         return Utils::getJsonAttr($this->options, $_key, $_default);
     }
 
-    public function setOptions($_key, $_value) {
+    public function setOptions($_key, $_value)
+    {
         $options = Utils::setJsonAttr($this->options, $_key, $_value);
-        $this->_changed = Utils::attrChanged($this->_changed,$this->options,$options);
+        $this->_changed = Utils::attrChanged($this->_changed, $this->options, $options);
         $this->options = $options;
         return $this;
     }
 
-    public function getRights($_key = '', $_default = '') {
+    public function getRights($_key = '', $_default = '')
+    {
         return Utils::getJsonAttr($this->rights, $_key, $_default);
     }
 
-    public function setRights($_key, $_value) {
+    public function setRights($_key, $_value)
+    {
         $rights = Utils::setJsonAttr($this->rights, $_key, $_value);
-        $this->_changed = Utils::attrChanged($this->_changed,$this->rights,$rights);
+        $this->_changed = Utils::attrChanged($this->_changed, $this->rights, $rights);
         $this->rights = $rights;
         return $this;
     }
 
-    public function getEnable() {
+    public function getEnable()
+    {
         return $this->enable;
     }
 
-    public function setEnable($_enable) {
-        $this->_changed = Utils::attrChanged($this->_changed,$this->enable,$_enable);
+    public function setEnable($_enable)
+    {
+        $this->_changed = Utils::attrChanged($this->_changed, $this->enable, $_enable);
         $this->enable = $_enable;
         return $this;
     }
 
-    public function getHash() {
+    public function getHash()
+    {
         if ($this->hash == '' && $this->id != '') {
             $hash = ConfigManager::genKey();
             while (is_object(UserManager::byHash($hash))) {
@@ -220,27 +241,32 @@ class User
         return $this->hash;
     }
 
-    public function setHash($_hash) {
-        $this->_changed = Utils::attrChanged($this->_changed,$this->hash,$_hash);
+    public function setHash($_hash)
+    {
+        $this->_changed = Utils::attrChanged($this->_changed, $this->hash, $_hash);
         $this->hash = $_hash;
         return $this;
     }
 
-    public function getProfils() {
+    public function getProfils()
+    {
         return $this->profils;
     }
 
-    public function setProfils($_profils) {
-        $this->_changed = Utils::attrChanged($this->_changed,$this->profils,$_profils);
+    public function setProfils($_profils)
+    {
+        $this->_changed = Utils::attrChanged($this->_changed, $this->profils, $_profils);
         $this->profils = $_profils;
         return $this;
     }
 
-    public function getChanged() {
+    public function getChanged()
+    {
         return $this->_changed;
     }
 
-    public function setChanged($_changed) {
+    public function setChanged($_changed)
+    {
         $this->_changed = $_changed;
         return $this;
     }
